+++ src/Esp32BaseConfig.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#ifndef ESP32BASE_CONFIG_H_COMPAT
-#define ESP32BASE_CONFIG_H_COMPAT
-#include "core/Esp32BaseConfig.h"
-#endif