+++ src/Esp32BaseLog.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#ifndef ESP32BASE_LOG_H_COMPAT
-#define ESP32BASE_LOG_H_COMPAT
-#include "core/Esp32BaseLog.h"
-#endif